ValidatorGrouping rework, Phase 1
Low-hanging fruit again. `ValidatorGrouping` now (mostly) doesn't care
*what* validators are in the entity's `Validators` namespace; it builds
a Hash of them all.

RSpec: 781 examples, 0 failures, 1 pending
RuboCop: 254 files inspected, no offenses detected
