+++ spec/entities/post/validators/image_url_spec.rb
@@ -1,16 +1,16 @@
 
 require 'spec_helper'
 
-require 'post/image_url_validator'
+require 'post/validators/image_url'
 
-require_relative 'shared/invalid_without_valid_body'
-require_relative 'shared/valid_with_valid_image_url'
-require_relative 'shared/valid_with_no_errors'
-require_relative 'shared/an_image_url_attribute'
+require_relative '../shared/invalid_without_valid_body'
+require_relative '../shared/valid_with_valid_image_url'
+require_relative '../shared/valid_with_no_errors'
+require_relative '../shared/an_image_url_attribute'
 
 # Namespace containing all application-defined entities.
 module Entity
-  describe Post::ImageUrlValidator do
+  describe Post::Validators::ImageUrl do
     let(:attributes) { FancyOpenStruct.new body: body, image_url: image_url }
     let(:obj) { described_class.new attributes }
 
@@ -20,5 +20,5 @@ module Entity
       other_shared = 'it is invalid without a valid body'
       it_behaves_like 'an image URL attribute', other_shared
     end # context 'when no body is specified in the attributes'
-  end # describe Post::ImageUrlValidator
+  end # describe Post::Validators::ImageUrl
 end
